fix(core): preserve repository planner boundary (#1110)
